requirements.txt
fastapi
uvicorn
pydantic
masonite-orm
python-dotenv
psycopg2-binary